Add Microsoft.Extensions.ServiceDiscovery.Yarp
