--- shared/Yarp.Extensions.Resilience/ServiceDiscovery/IServiceDiscoveryDestinationResolver.cs
@@ -1,9 +1,0 @@
-using Yarp.ReverseProxy.ServiceDiscovery;
-
-namespace Yarp.Extensions.Resilience.ServiceDiscovery;
-
-public interface IServiceDiscoveryDestinationResolver
-{
-    ValueTask<ResolvedDestinationCollection> ResolveDestinationsAsync(string serviceName,
-        CancellationToken cancellationToken);
-}
