--- shared/Yarp.Extensions.Resilience/ServiceDiscovery/ServiceDiscoveryOptions.cs
@@ -1,6 +1,0 @@
-namespace Yarp.Extensions.Resilience.ServiceDiscovery;
-
-public class ServiceDiscoveryOptions
-{
-    public TimeSpan? RefreshPeriod { get; set; } = TimeSpan.FromMinutes(5);
-}
