+++ src/Sail.Proxy/Sail.Proxy.csproj
@@ -12,4 +12,8 @@
       <ProjectReference Include="..\Sail.Core\Sail.Core.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.ServiceDiscovery.Yarp" />
+    </ItemGroup>
+
 </Project>
