+++ src/Sail/Apis/RouteApi.cs
@@ -1,8 +1,7 @@
 using Microsoft.AspNetCore.Http.HttpResults;
-using Sail.Core.Entities;
 using Sail.Extensions;
+using Sail.Models.Route;
 using Sail.Services;
-using RouteMatch = Sail.Api.V1.RouteMatch;
 
 namespace Sail.Apis;
 
@@ -59,20 +58,3 @@ public static class RouteApi
         );
     }
 }
-
-public record RouteRequest(Guid ClusterId, string Name, RouteMatchRequest Match);
-
-public record RouteMatchRequest(
-    string Path,
-    List<string>? Methods,
-    List<string>? Hosts,
-    List<QueryParameterRequest>? ParameterRequests,
-    List<RouteHeaderRequest>? RouteHeaders);
-
-public record QueryParameterRequest(
-    string Name,
-    List<string> Values,
-    QueryParameterMatchMode Mode,
-    bool IsCaseSensitive);
-
-public record RouteHeaderRequest(string Name, List<string> Values, HeaderMatchMode Mode, bool IsCaseSensitive);
