+++ src/Sail/Grpc/RouteGrpcService.cs
@@ -8,7 +8,6 @@ using Sail.Storage.MongoDB.Extensions;
 using Route = Sail.Core.Entities.Route;
 using RouteResponse = Sail.Api.V1.Route;
 
-
 namespace Sail.Grpc;
 
 public class RouteGrpcService(SailContext dbContext, IRouteStore routeStore) : RouteService.RouteServiceBase
@@ -85,10 +84,10 @@ public class RouteGrpcService(SailContext dbContext, IRouteStore routeStore) : R
             Order = route.Order,
             ClusterId = route.ClusterId?.ToString(),
             CorsPolicy = route.CorsPolicy,
-            Timeout = -1,
+            Timeout = route.Timeout?.ToString("hh:mm:ss"),
             TimeoutPolicy = route.TimeoutPolicy,
             AuthorizationPolicy = route.AuthorizationPolicy,
-            MaxRequestBodySize = route.MaxRequestBodySize ?? -1,
+            MaxRequestBodySize = route.MaxRequestBodySize,
             RateLimiterPolicy = route.RateLimiterPolicy
         };
     }
