+++ src/Sail/Models/Route/QueryParameterRequest.cs
@@ -1,0 +1,9 @@
+using Sail.Core.Entities;
+
+namespace Sail.Models.Route;
+
+public record QueryParameterRequest(
+    string Name,
+    List<string> Values,
+    QueryParameterMatchMode Mode,
+    bool IsCaseSensitive);
