+++ src/Sail/Models/Route/RouteMatchRequest.cs
@@ -1,0 +1,8 @@
+namespace Sail.Models.Route;
+
+public record RouteMatchRequest(
+    string Path,
+    List<string>? Methods,
+    List<string>? Hosts,
+    List<QueryParameterRequest>? ParameterRequests,
+    List<RouteHeaderRequest>? RouteHeaders);
