+++ src/Sail/Models/Route/RouteRequest.cs
@@ -1,0 +1,3 @@
+namespace Sail.Models.Route;
+
+public record RouteRequest(Guid ClusterId, string Name, RouteMatchRequest Match);
