+++ src/Sail/Sail.csproj
@@ -15,6 +15,7 @@
     
     <ItemGroup>
         <ProjectReference Include="..\..\shared\ServiceDefaults\ServiceDefaults.csproj" />
+        <ProjectReference Include="..\Sail.Compass\Sail.Compass.csproj" />
         <ProjectReference Include="..\Sail.Storage.MongoDB\Sail.Storage.MongoDB.csproj" />
     </ItemGroup>
