+++ src/Sail/Services/ClusterService.cs
@@ -91,7 +91,7 @@ public record ClusterVm
 
 public record DestinationVm
 {
-    public string Address { get; init; }
+    public string? Address { get; init; }
     public string? Health { get; init; }
     public string? Host { get; init; }
 }
